test: merge whois integration tests into unit tests
Merged TestWhois_PreservesAllHeaders and TestWhois_HandlesPartialResponse
from the integration test file into the unit test file. These tests
verify the same functionality without needing a separate integration
test file.

Removed whois_integration_test.go as all test coverage is maintained
in the consolidated unit test file.
